Tried fixing RFLAG issue not working
Attempt at fixing the problem of only working on the 3 copies if RFLAG but not working.  Changed PUT for copy1 if rflag but parseHeader needs to be changed instead b/c content length is pointing at main file.

Other issues relating to the above is commented as TODO: issues

isMAIN arg for copyFiles is working and will be copied to main.